Revert "Feature/news styling"
