+++ client/Public/index.html
@@ -7,7 +7,6 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Jade City Records</title>
   <link href="../src/style.css" rel="stylesheet" />
-  <link rel="icon" type="image/svg+xml" href="./img/JC-Logo.svg">
   <!-- Belleza Font -->
   <link rel="preconnect" href="https://fonts.googleapis.com">
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
