+++ client/src/Pages/About.jsx
@@ -2,29 +2,35 @@ import React from 'react'
 
 const About = () => {
     return (
-        <section id="about">
-          <div className="aboutContainer">
-            <p className="welcome">
-                Welcome to
-                <span className="records">JADE CITY,</span>
-            </p>
-            <h2 className="City">the City of Mages</h2>
-            <p>
-                — your new home for ethereal melodic bass. With
-                immersive songwriting, enchanting vocals, and
-                spellbinding production, the mages of Jade City are
-                bringing a new magic to the world of bass music. More
-                than just a record label, Jade City is a thriving
-                community of music creators and enthusiasts who gather
-                under the banner of creativity and kindness.
-            </p>
-            <p>
-                Come and discover a realm where every note is a spell,
-                every lyric a story, and every gathering a celebration
-                of the magic of friendship.
-            </p>
-            </div>
-        </section>
+        <>
+            <div className="gradient"></div>
+            <section id="about">
+                <p className="welcome">
+                    Welcome to...
+                    <span className="records">JADE CITY</span>
+                </p>
+                <h2>Where our... Lorem ipsum dolor sit amet!</h2>
+                <p>
+                    Here we value ... Lorem ipsum dolor sit amet
+                    consectetur adipisicing elit. Eum vel facilis
+                    perspiciatis deserunt temporibus minus esse atque
+                    voluptas! Dolores vel, neque aspernatur placeat
+                    excepturi sit temporibus, sunt vitae doloremque ex
+                    delectus consectetur cumque quos corrupti eos hic
+                    et adipisci dolore distinctio iure cum laudantium!
+                    Ad sequi laudantium assumenda ipsa. Sed rerum ipsa
+                    accusantium tenetur! Delectus magni incidunt
+                    cumque temporibus accusamus, sit cupiditate velit
+                    iusto, nam hic debitis odio! Voluptatem quo iste
+                    aliquam, dolorum possimus a asperiores culpa autem
+                    laborum, harum quam sint temporibus nam ex vitae
+                    sit placeat corrupti pariatur quisquam perferendis
+                    soluta aperiam. Cumque minus ad dolor nihil
+                    impedit?
+                </p>
+            </section>
+            <div className="gradient"></div>
+        </>
     )
 }
