+++ client/src/Pages/Artists.jsx
@@ -1,6 +1,6 @@
 // src/components/ArtistSection.js
 import React, { useEffect, useState } from 'react';
-import ArtistCard from '../components/ArtistCard/ArtistCard.jsx'; // Adjust the path if necessary
+import ArtistCard from '../components/ArtistCard'; // Adjust the path if necessary
 
 const Artists = (props) => {
