--- client/src/Pages/Contact/Contact.css
@@ -1,45 +1,0 @@
-/* contact section */
-fieldset {
-    border: 1px var(--font-color); /* Removes the default border */
-    padding: 0 0 1rem 0; /* Removes any padding */
-    font-size: 1em; /* You can adjust this to your needs */
-}
-
-#contact {
-    padding: 4rem 1rem;
-    margin: 0 auto;
-    max-width: 1200px;
-}
-
-#contact h2 {
-  font-size: 2.5rem;
-  margin-bottom: 1.5rem;
-}
-
-#contact .demoForm {
-    max-width: auto;
-    margin: 1rem 4rem;
-    padding: .5rem;
-    text-align: center;
-    background-color: var(--secondary);
-    font-family: var(--primary-font);
-    border-radius: var(--comp-b-radius);
-}
-
-#contactHeading {
-    margin: 0;
-    padding: .5rem 0 1rem 0;
-}
-
-#contact p {
-    font-size: 1rem;
-    margin: 0;
-    padding: .5rem;
-}
-
-.additionalInfo {
-    display: flex;
-    flex-wrap: row wrap;
-    align-items: center;
-    justify-content: center;
-}
