--- client/src/Pages/Shop/Shop.css
@@ -1,20 +1,0 @@
-#shop {
-    padding: 4rem 1rem;
-    margin: 0 auto;
-    max-width: 1200px;
-}
-
-#shop h2 {
-  font-size: 2.5rem;
-  margin-bottom: .5rem;
-}
-
-.shop-list {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-}
-
-.loading {
-    margin: auto;
-}
