--- client/src/components/ArtistCard/artistCard.css
@@ -1,25 +1,0 @@
-.artist-card {
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    align-items: center; /* Centers the content horizontally */
-    background-color: var(--secondary);
-    padding: 1rem;
-    margin: 1px;
-    width: 100%;
-    max-width: 300px;
-    min-height: 450px;
-    border-radius: var(--comp-b-radius);
-    text-align: center;
-}
-
-.artist-card img {
-    margin-bottom: auto;
-    border-radius: 8px;
-}
-
-.artist-card .name {
-    margin: 0;
-    margin-top: auto;
-    padding-bottom: 1rem;
-}
