+++ client/src/components/ArtistCardLarge.jsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import { useLocation } from 'react-router-dom';
-import styles from './artistCardLarge.css'
 
 
 function ArtistCardLarge() {
@@ -9,22 +8,20 @@ function ArtistCardLarge() {
     const { name, bio, imgUrl, inspiration, spotifyUrl } = location.state; 
 
     return (
-        <div className="artist-card-large">
+        <div className="artist-card-large" onClick={() => handleArtistClick()}>
             <img
                 className="artist-img-large"
                 src={imgUrl}
                 alt={`Photo of ${name}`}
             />
-            <h3>{name}</h3>
             <div className="artist-info-large">
+                <h3>{name}</h3>
                 <p>{bio}</p>
                 <p>
                     <strong>Inspiration:</strong> {inspiration}
                 </p>
             </div>
-            <div className="spotify">
             <iframe src={spotifyUrl} width="100%" height="352" frameBorder="0" allow="autoplay; clipboard-write; encrypted-media; fullscreen; picture-in-picture" loading="lazy"></iframe>
-            </div>
         </div>
     )
